Moved local scheduler test to a separate file
